Moved Nano-specific console commands to nano folder
