+++ main/coins/nano/console.h
@@ -1,0 +1,6 @@
+#ifndef __JOLT_NANO_CONSOLE_H__
+#define __JOLT_NANO_CONSOLE_H__
+
+void console_nano_register();
+
+#endif
