+++ main/console.c
@@ -1,7 +1,4 @@
 #include "menu8g2.h"
-#include "vault.h"
-#include "globals.h"
-#include "loading.h"
 #include "sodium.h"
 #include "freertos/FreeRTOS.h"
 #include "freertos/queue.h"
@@ -18,6 +15,9 @@
 #include "nano_parse.h"
 #include "globals.h"
 #include "console.h"
+#include "vault.h"
+
+#include "coins/nano/console.h"
 
 static const char* TAG = "console";
 
@@ -27,78 +27,6 @@ volatile TaskHandle_t console_h = NULL;
 static int free_mem(int argc, char** argv) {
     printf("Free: %d bytes\n", esp_get_free_heap_size());
     return 0;
-}
-
-static int nano_count(int argc, char** argv) {
-    printf("nano_count\n");
-    uint32_t block_count = nanoparse_lws_block_count();
-    printf("Returned: %d\n", block_count);
-    return 0;
-}
-
-static int nano_balance(int argc, char** argv) {
-    size_t disp_buffer_size = 8 * u8g2_GetBufferTileHeight(&u8g2) *
-            u8g2_GetBufferTileWidth(&u8g2);
-    uint8_t *old_disp_buffer = malloc(disp_buffer_size);
-    memcpy(old_disp_buffer, u8g2_GetBufferPtr(&u8g2), disp_buffer_size);
-
-    vault_rpc_t rpc;
-    double display_amount;
-    
-    /******************
-     * Get My Address *
-     ******************/
-    nvs_handle nvs_h;
-    init_nvm_namespace(&nvs_h, "nano");
-    if(ESP_OK != nvs_get_u32(nvs_h, "index", &(rpc.nano_public_key.index))){
-        rpc.nano_public_key.index = 0;
-    }
-    nvs_close(nvs_h);
-    
-    rpc.type = NANO_PUBLIC_KEY;
-    if(vault_rpc(&rpc) != RPC_SUCCESS){
-        printf("User cancelled.\n");
-        display_amount = -1;
-        goto exit;
-    }
-    char address[ADDRESS_BUF_LEN];
-    nl_public_to_address(address, sizeof(address),
-            rpc.nano_public_key.block.account);
-    printf("Address: %s\n", address);
-    
-    /********************************************
-     * Get My Account's Frontier Block *
-     ********************************************/
-    // Assumes State Blocks Only
-    // Outcome:
-    //     * frontier_hash, frontier_block
-    nl_block_t frontier_block;
-    nl_block_init(&frontier_block);
-    memcpy(frontier_block.account, rpc.nano_public_key.block.account, BIN_256);
-    
-    switch( nanoparse_lws_frontier_block(&frontier_block) ){
-        case E_SUCCESS:
-            if( E_SUCCESS != nl_mpi_to_nano_double(&(frontier_block.balance),
-                                                   &display_amount) ){
-                display_amount = -1;
-            }
-            break;
-        default:
-            display_amount = 0;
-            break;
-    }
-
-    if( display_amount >= 0 ){
-        printf("Balance: %0.4lf Nano\n", display_amount);
-    }
-    else{
-        printf("Error\n");
-    }
-
-    exit:
-        memcpy(u8g2_GetBufferPtr(&u8g2), old_disp_buffer, disp_buffer_size);
-        free(old_disp_buffer);
-        return display_amount;
 }
 
 void console_task() {
@@ -192,28 +120,15 @@ static void console_register_commands(){
     esp_console_register_help_command();
 
     cmd = (esp_console_cmd_t) {
-        .command = "nano_count",
-        .help = "Get the current Nano block count",
-        .hint = NULL,
-        .func = &nano_count,
-    };
-    ESP_ERROR_CHECK( esp_console_cmd_register(&cmd) );
-
-    cmd = (esp_console_cmd_t) {
-        .command = "nano_balance",
-        .help = "Get the current Nano Balance",
-        .hint = NULL,
-        .func = &nano_balance,
-    };
-    ESP_ERROR_CHECK( esp_console_cmd_register(&cmd) );
-
-    cmd = (esp_console_cmd_t) {
         .command = "free",
         .help = "Get the total size of heap memory available",
         .hint = NULL,
         .func = &free_mem,
     };
     ESP_ERROR_CHECK( esp_console_cmd_register(&cmd) );
+
+    /* Register Coin Specific Commands */
+    console_nano_register();
 }
 
 void initialize_console() {
