+++ main/console.h
@@ -3,7 +3,6 @@
 
 void initialize_console();
 void menu_console(menu8g2_t *prev);
-void backend_task(void *backend_in);
 volatile TaskHandle_t *start_console();
 
 #endif
